[refactor] widgets 폴더 내 파일들 용도별 폴더 분리 #362
