+++ frontend/lib/screen/add_company_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/widgets/button/bottom_text_button.dart';
 import 'package:frontend/widgets/dialog/one_button_dialog.dart';
-import 'package:frontend/widgets/iconed_textfield.dart';
+import 'package:frontend/widgets/textfield/iconed_textfield.dart';
 import 'package:frontend/service/api_service.dart';
-import 'package:frontend/widgets/top_appbar.dart';
+import 'package:frontend/widgets/bar/top_appbar.dart';
 
 class AddCompanyScreen extends StatefulWidget {
   const AddCompanyScreen({super.key});
@@ -91,7 +91,7 @@ class _AddCompanyScreenState extends State<AddCompanyScreen> {
       showDialog(
         context: context,
         builder: (BuildContext context) {
-          return AlertDialog(
+          return const AlertDialog(
             content: OneButtonDialog(
               first: '회사명을 입력해주세요.',
             ),
@@ -104,7 +104,7 @@ class _AddCompanyScreenState extends State<AddCompanyScreen> {
       showDialog(
         context: context,
         builder: (BuildContext context) {
-          return AlertDialog(
+          return const AlertDialog(
             content: OneButtonDialog(
               first: '사업자 등록번호를 입력해주세요.',
             ),
@@ -117,7 +117,7 @@ class _AddCompanyScreenState extends State<AddCompanyScreen> {
       showDialog(
         context: context,
         builder: (BuildContext context) {
-          return AlertDialog(
+          return const AlertDialog(
             content: OneButtonDialog(
               first: '사내메일 도메인을 입력해주세요.',
             ),
@@ -131,7 +131,7 @@ class _AddCompanyScreenState extends State<AddCompanyScreen> {
       showDialog(
         context: context,
         builder: (BuildContext context) {
-          return AlertDialog(
+          return const AlertDialog(
             content: OneButtonDialog(
               first: '도메인 형식이 올바르지 않습니다.',
             ),
