+++ frontend/lib/screen/position_select_screen.dart
@@ -4,7 +4,7 @@ import 'package:frontend/screen/edit_profile_screen.dart';
 import 'package:frontend/service/api_service.dart';
 import 'package:frontend/widgets/button/bottom_text_button.dart';
 import 'package:frontend/widgets/dialog/one_button_dialog.dart';
-import 'package:frontend/widgets/top_appbar.dart';
+import 'package:frontend/widgets/bar/top_appbar.dart';
 import 'package:provider/provider.dart';
 
 class PositionSelectScreen extends StatefulWidget {
@@ -154,7 +154,7 @@ class PositionSelectScreenState extends State<PositionSelectScreen> {
       showDialog(
         context: context,
         builder: (BuildContext context) {
-          return AlertDialog(
+          return const AlertDialog(
             content: OneButtonDialog(
               first: '직무 리스트를 불러올 수 없습니다.',
             ),
