+++ frontend/lib/widgets/thermometer/big_thermometer.dart
