+++ frontend/lib/widgets/thermometer/thermometer.dart
