+++ frontend/lib/widgets/user/user_item.dart
@@ -6,9 +6,9 @@ import 'package:frontend/screen/chat_screen.dart';
 import 'package:frontend/screen/matching_screen.dart';
 import 'package:frontend/service/api_service.dart';
 import 'package:frontend/service/auto_offline_service.dart';
-import 'package:frontend/widgets/user_details_modal.dart';
+import 'package:frontend/widgets/user/user_details_modal.dart';
 import 'package:frontend/widgets/choose_purpose.dart';
-import 'package:frontend/widgets/user_details.dart';
+import 'package:frontend/widgets/user/user_details.dart';
 import 'package:frontend/widgets/color_text_container.dart';
 import 'package:frontend/widgets/button/bottom_two_buttons.dart';
 import 'package:frontend/widgets/profile_img.dart';
